requirements.txt
pandas 
streamlit 
requests 
seaborn 
matplotlib 
sqlite3
python-dotenv==0.19.0
virtualenv==20.13.0

# To create and activate virtual environment:
# On Windows:
# python -m venv venv
# .\venv\Scripts\activate
#
# On Mac/Linux:
# python3 -m venv venv
# source venv/bin/activate
#
# Then install requirements:
# pip install -r requirements.txt

